Implemented setPaused, setRun, setStep core functions
Added functions to allow setting the pause, run, and step states.

The run state is false by default, and when starting emulation it is set to true. When emulation is to permanently end, run is set false so terminate the emulation loop.

The pause state allows skipping the emulation functions each loop so that it does not proceed further until the pause state is set false again.

The step state allows overriding the pause state once. This is used for debugging.