Added basic readCPURam function
This function will be used for reading a single byte of data from cpu ram. It will also do special functionality when reading from specific registers.

Some of this will not be implemented until the PPU class (and it's necessary functions) are implemented.

For now, it can read the joypad register and read from cpu ram.