Added basic writeCPURam function
This function handles writes to cpu ram. It will apply the cycle cost to make the write as well as set the appropriate value at the address given, unless the address is a special address in which may result in other functionality.

For now, that only encompasses writing to the joypad write register, but will be expanded when the PPU is implemented.