Added directCPURamWrite to bypass effecting special registers
Wehn the normal writeCPURam function is called, it may not write to the cpu ram if the address given is a special register.

This means that if overwriting the cpu ram occurs, one may accidentally unintentionally modify registers unrelated to the raw ram data.

The directCPURamWrite has been added to allow writing directly to the array without effecting the emulation.