Fixed issues with reversing tile index arrays
In three different functions an array may need to be reversed. There were two issues  with the original logic, one was an off by one error where the array count was not subtracted by 1 before being used to index the array.

The other issue was looping over the length of the array instead of half of the length of the array. If it loops over the whole array then each element will actually get swapped twice, resulting in the original array. These have both been fixed.