Fixed statusRegister not being initialized properly
Previously the statusRegister did not initialize with it's intended value. This is because it uses a bitwise or to set two flags enabled.

However, the value for these flags were not defined until after the or operation occurred. Now they are ordered so that they will be initialized first so they can be used by the statusRegister.