The rom header is now parsed upon rom loading
In order to get information about the rom file required for emulation like what kind of rom mapper it uses the header of the file must be read in.

This is now done when the file is loaded in initially.

Additionally emulation is now paused before loading it in so that if a previous game is being executed there will be no writes to memory that is being overwritten with the new rom load.

It is then unpaused and emulation is set to run once it is complete.