Added Mapper class for handling writes to Mapper addresses
Mapper emulation requires handling writes to specific addresses to control aspects of the mapper. This can be different for each mapper. This class will allow checking if the address corresponds to any of the current emulated mapper's special addresses, and if so, process the write.

For now, the only mapper supported is NROM which does not require this. However, this work is added in order to handle the addition of MMC1, MMC3, UNROM, and CNROM which do.