Removed unused time fragment shader variable
The float variable was unintentionally added and was a left over from some shader tests. This went unnoticed until it was flagged by the MingW compiler as it did not define a precision.

If this variable were to return that change would be made, however it is being removed since it was not intended to be removed.