Open Palette dialog now always starts at application directory
Previously if the palette was selected before a rom file it would open at the application directory. However, if a dialog was opened prior it would start at whichever directory it selected.

This was not ideal, as palettes are likely to be stored with the emulator while rom backup files are likely in a universal folder to be used by multiple emulators.

This way it will always be consistent.