Fixed OpenGL Canvas Aspect Ratio
Previously the 3D vertices were not arranged optimally to exactly match the NES aspect ratio. While it was generally close, the difference resulted in some pixel lines would either not render or render double due to it stretching or contracting to fit the frame on the incorrectly shaped 3D quad.

This was most obvious when scrolling such as in 2D platforming games.

This has been carefully adjusted to render the screen in perfect aspect ratio.