Fixed 2 lines being cut off and adjust window size
In a previous change the aspect ratio was corrected to render pixel perfect. This however needed to be accompanied by two changes in order to work with the QLayout.

First, the 3D quad needed to be moved up two pixels so they would not get cut off.

Secondly, the application window had to be shrinked by two pixels on it's width and height.
2 pixels were originally added to account for the QLayout, however are no longer required. In fact, the 2 pixels will actually make the QLayout not display the aspect ratio correctly since it will attempt to fit the extra 2 pixels.

This has been fixed.