Added Window Scale settings to menu for perfect aspect ratio resizing
The window can be resized and scaled however the user likes. However, it can be very difficult if not impossible to manually scale to a new resolution that is in the same aspect ratio. If it is not in the same aspect ratio the graphics may not display pixel perfect.

In order to make it easy to scale the emulator up to 2 times, 3 times, 4 times, or back to the normal resolution without issue, a set of menu options have been added. Now one can switch to any of these window scale sizes by going to File->Settings->Scale and selecting one of the 4 options.

This runs alongside being able to scale manually, so both options are available for convenience.