Added support for loading save files with the same name as loaded rom
When a rom file is loaded into the emulator, it will now check for a file with the same name but with the .sav extension. If this file exists, the byte data contents of the file will be loaded into cpu ram to be accessed by emulation.

This will allow playing the game with existing save data and continuing where they player left off if the game has save data support.