Loading a rom while another is being played now works correctly
Previously if one wished to switch to another rom after launching another one, it would be best to restart the emulator and load the new rom.

This was because the emulation state did not get reset when loading a new rom. Instead the new rom was loaded in, then emulation would start. This means the cpu and ppu would be in significantly different states than it would when emulation begins. The programming of the games often depend on the system being in a certain state when it starts off, so the emulation of the newly loaded rom would act strangely.

Now when a new rom is loaded, if one was loaded prior, it will reset each of the components to have their original state again before loading and playing the new rom.