Fixed No Input after loading a second rom file
Previously work was done to reset the emulation state to it's original state when a rom file is loaded so that emulation would work as intended when loading a new rom after one has already been loaded in.

This worked well, however, due to an unintended mistake by setting the joy1Connected variable to false instead of true, no keyboard input would work once the next rom has been loaded. This is now fixed.