MMC1 and MMC3 mappers now get reset on new rom load
Previously work was done so that after emulation occurs and a second rom is loaded in to be emulated all of the emulation classes would reset their state to how it was originally for a clean slate to emulate from.

However, the MMC1 and MMC3 mappers did not get included in this change. They will be reset as well now.