Remove unused direct glMatrixMode call blocking MingW compilation
This call to glMatrixMode goes unused as glMatrixMode is not used in modern opengl and this program does not used a fixed state rendering pipeline. The equivalent would be performed in the shader.

This was causing issues with compiling in MingW and VS since Qt does not implement an equivalent function. When compiling on Windows it would require linking with opengl directly instead of depending on Qt. Since this call is not needed and complicates compilation it is being removed.