Removed unused Binary helper class
Originally a Binary class was created as a helper class to implement binary rotation. However, the user case for this (the rotate assembly instructions) do not rotate directly. Instead the rotation is handled as a shift as well as setting flags for the CPU to be aware of the bit that was shifted off.

The bit is not directly appended  to the opposite side. Therefore this class is not needed and is being removed.