Handle second bank for 16kb CNROM games
If a CNROM has 32kb of program rom data the first 16kb will be be read into CPU Ram at 0x8000. The second half will be at 0xC000.

However, if the program rom is only 16 kb it will be at 0x8000 and the same data will be mirrored at 0xC000. This fixes issues where the rest vector is 0x0000 due to the mirrored data not being loaded in.