Updated colors of initial main window screen
Updated the initial background of the main window to match that of the C# version of GameBeak. The C++ version of GameBeak has also been updated to use this style for consistency.

This is simply 5 rows of purple on the top, followed by the rest of the screen being pink.