Added MenuBar option to create a .sav file
If a game supports using save ram to hold save data for a game, it is now possible to save that ram data to a file by pressing the create save file option under the Status menu bar.

Previously this was possible by pressing F1, but now there is a way to do it via the UI.

If a rom is not loaded or the rom does not support save ram then this option will be disabled until a rom that does support it is loaded.